requirements.txt
google-generativeai>=0.4.1
requests>=2.31.0
python-dotenv>=1.0.1
